+++ apps/fluent-tester/src/TestComponents/RadioGroupExperimental/RadioGroupExperimentalTest.tsx
@@ -4,6 +4,7 @@ import { DefaultRadioGroup } from './DefaultRadioGroup';
 import { E2ERadioGroupExperimentalTest } from './RadioGroupExperimentalE2ETest';
 import { RequiredRadioGroup } from './RequiredRadioGroup';
 import { DisabledRadioGroup } from './DisabledRadioGroup';
+import { SubtextRadioGroup } from './SubtextRadioGroup';
 import { Test, TestSection, PlatformStatus } from '../Test';
 
 const radioGroupExperimentalSections: TestSection[] = [
@@ -21,6 +22,11 @@ const radioGroupExperimentalSections: TestSection[] = [
     name: 'Required RadioGroup',
     testID: RADIO_GROUP_EXPERIMENTAL_TESTPAGE,
     component: RequiredRadioGroup,
+  },
+  {
+    name: 'RadioGroup with Label Subtext',
+    testID: RADIO_GROUP_EXPERIMENTAL_TESTPAGE,
+    component: SubtextRadioGroup,
   },
   {
     name: 'RadioGroup for E2E Testing',