+++ packages/deprecated/theming-react-native/src/ThemeProvider.types.ts
@@ -2,6 +2,9 @@
 
 import { ThemeRegistry } from './Theme.types';
 
+/**
+ * @deprecated
+ */
 export interface IThemeProviderProps extends React.PropsWithChildren<{}> {
   theme?: string;
   registry?: ThemeRegistry;