+++ packages/deprecated/theming-ramp/CHANGELOG.md
@@ -1,8 +1,41 @@
 # Change Log - @uifabricshared/theming-ramp
 
-This log was last generated on Wed, 21 Sep 2022 21:00:40 GMT and should not be manually modified.
+This log was last generated on Wed, 12 Oct 2022 21:54:15 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.17.6
+
+Wed, 12 Oct 2022 21:54:15 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/theme-types to v0.22.0
+
+## 0.17.5
+
+Tue, 11 Oct 2022 22:41:44 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/theme-types to v0.21.0
+
+## 0.17.4
+
+Fri, 30 Sep 2022 08:04:44 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/theme-types to v0.20.0
+
+## 0.17.3
+
+Fri, 30 Sep 2022 00:54:37 GMT
+
+### Patches
+
+- Bump @uifabricshared/foundation-settings to v0.12.1
+- Bump @fluentui-react-native/theme-types to v0.19.2
 
 ## 0.17.2